Fix week navigation
